# internal/grpc/github.com/PluT00/mobile-otp/api/mobile-otp/mobile-otp.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mobile-otp/mobile-otp.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MobileOTP"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/get-otp": {
      "post": {
        "summary": "GetOTP - send previously generated OTP to user",
        "operationId": "MobileOTP_GetOTP",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/otpgrpcGetOTPResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/otpgrpcGetOTPRequest"
            }
          }
        ],
        "tags": [
          "MobileOTP"
        ]
      }
    },
    "/sync-keys": {
      "post": {
        "summary": "SyncKeys - sync ecdh public keys",
        "operationId": "MobileOTP_SyncKeys",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/otpgrpcSyncKeysResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/otpgrpcSyncKeysRequest"
            }
          }
        ],
        "tags": [
          "MobileOTP"
        ]
      }
    }
  },
  "definitions": {
    "otpgrpcGetOTPRequest": {
      "type": "object",
      "properties": {
        "encryptedJwt": {
          "type": "string",
          "title": "Encrypted JWT, must be with mobile=true flag"
        },
        "nonce": {
          "type": "integer",
          "format": "int32",
          "title": "Random number sent previously in SyncKeys request"
        }
      }
    },
    "otpgrpcGetOTPResponse": {
      "type": "object",
      "properties": {
        "encryptedOtp": {
          "type": "string",
          "title": "Encrypted OTP (Base64)"
        },
        "success": {
          "type": "boolean",
          "title": "Success flag; false if there is no request to generate OTP for that user"
        }
      }
    },
    "otpgrpcNeedOTPResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "otpgrpcOTP": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        }
      }
    },
    "otpgrpcSyncKeysRequest": {
      "type": "object",
      "properties": {
        "publicKey": {
          "type": "string",
          "title": "Client public key (Base64)"
        },
        "nonce": {
          "type": "integer",
          "format": "int32",
          "title": "Random number to find public key later in GetOTP request"
        }
      }
    },
    "otpgrpcSyncKeysResponse": {
      "type": "object",
      "properties": {
        "publicKey": {
          "type": "string",
          "title": "Server public key (Base64)"
        }
      }
    },
    "otpgrpcValidateOTPResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}